Extract folder functions to utils
